Type hints for constants.

+++ core/src/config.jl
@@ -138,11 +138,11 @@ end
     autodiff::Bool = true
     evaporate_mass::Bool = true
     # Universal depth at which the low storage factor kicks in
-    low_storage_depth = 0.1
+    low_storage_depth::Float64 = 0.1
     # Universal reduction factor threshold for the minimum upstream level of UserDemand nodes
-    user_demand_min_level_threshold = 0.1
+    user_demand_min_level_threshold::Float64 = 0.1
     # Universal reduction factor threshold for the level difference of Pump/Outlet and TabulatedRatingCurve nodes
-    flow_reduction_factor_threshold = 0.02
+    flow_reduction_factor_threshold::Float64 = 0.02
 end
 
 @option struct Interpolation <: TableOption
